requirements.txt
transformers~=4.41.2
flask~=3.0.3
PyYAML~=6.0.1